Crate gourd_lib

The architecture of our codebase, shared between wrapper and CLI.

Modules

config

A struct and related methods for global configuration, declaratively specifying experiments.

constants

Constant values.

error

The error handling for gourd.

experiment

The setup of an experiment.

file_system

Common file operations

measurement

Code shared between the wrapper and gourd.

network

Interactions with the network for fetching resources.

resources

Resource fetching helpers.

Macros

bailc

This is a shorthand for the anyhow::bail macro, now with context.

ctx

This is a shorthand for returning the context of a error.