gourd_lib

Module config

A struct and related methods for global configuration, declaratively specifying experiments.

Modules

fetching

Fetching for resources.

maps

Deserializer for the maps.

parameters

Deserializer for the paramters (grid search).

slurm

Slurm configuration.

Structs

Config

A config struct used throughout the gourd application.

FetchedResource

Fetch a remote resource

GitProgram

An algorithm fetched from a git repository.

InputSchema

TOML struct that can be used to provide inputs.

Label

A label that can be assigned to a job based on the afterscript output.

LocalOptions

Options for configuring the execution of gourd run local

Parameter

A parameter.

Regex

A wrapper around [regex_lite::Regex] to allow serde.

SubParameter

A subparameter.

UserInput

A pair of a path to an input and additional cli arguments.

UserProgram

A pair of a path to a binary and cli arguments.