gourd_lib

Module constants

Constant values.

Constants

AFTERSCRIPT_DEFAULT

The default path to the afterscript.

CMD_DOC_STYLE

Style of commands in doc messages

CMD_STYLE

Style of commands in help messages

EMPTY_ARGS

The default arguments for an input.

ERROR_STYLE

The styling for error messages.

GLOB_ESCAPE

The prefix which will cause an argument to be interpreted as a glob. Ensure matches:

GOURD_VERSION

The version name for Gourd! Ensure it matches:

HELP_STYLE

The styling for help messages.

INTERNAL_GLOB

The internal representation of globbed inputs.

INTERNAL_PARAMETER

The internal representation of paramtrized inputs.

INTERNAL_PREFIX

Internal representation for names parsed from config

INTERNAL_SCHEMA_INPUTS

The internal representation of inputs generated from a schema

LABEL_OVERLAP_DEFAULT

The default value of warning on label overlaps.

LOGO

The logo of the application.

MAIL_TYPE_VALID_OPTIONS

Possible values for Mail Type in slurm configuration

NAME_STYLE

The styling for the university name.

PARAGRAPH_STYLE

Grayed paragraphs of not very important text

PARAMETER_ESCAPE

The prefix which will cause an argument to be interpreted as a parameter. Ensure matches:

PATH_STYLE

Style of std::path::Paths and PathBufs

PLOT_SIZE

The size of the analysis output plots, in pixels.

POSTPROCESS_JOB_DEFAULT

The default postprocess job name.

PRIMARY_STYLE

The styling for the program name.

PROGRAM_RESOURCES_DEFAULT

The default value of resource limits for a program.

RERUN_LABEL_BY_DEFAULT

Do we assume by default that runs with custom labels are failed runs?

RERUN_LIST_PROMPT_CUTOFF

Maximal number of individual prompts that the user can be asked when trying to rerun

SCHEDULE_BAR_WIDTH

The length of the bar for scheduling.

SECONDARY_STYLE

The styling for the secondary text.

SHORTEN_STATUS_CUTOFF

The maximal amount of runs before status only shows a short summary.

SLURM_VERSIONS

Supported SLURM versions.

STATUS_REFRESH_PERIOD

The amount between refreshes of the status screen, in ms.

SUB_PARAMETER_ESCAPE

The prefix which will cause an argument to be interpreted as a subparameter. Ensure matches:

TASK_LIMIT

The maximal amount of tasks that gourd will schedule.

TERTIARY_STYLE

The styling for the tertiary text.

TRUNCATE_AFTERSCRIPT_OUTPUT

The max (bytes, lines) of afterscript output that will be displayed in gourd status of a run

WARNING_STYLE

The styling for warning messages.

WRAPPER_DEFAULT

The default path to the wrapper, that is, we assume gourd_wrapper is in $PATH.

Functions

style_from_fg

Create a style with a defined foreground color.