gourd_lib::error

Macro ctx

macro_rules! ctx {
    {$cause: expr,  $($arg_cause: expr)*; $help: expr, $($arg_help: tt)*} => { ... };
}

This is a shorthand for returning the context of a error.

This macro takes a context message, a help message and formats them.

A macro invocation like so:

ctx!([context], [context arg 1], [context arg 2], ...; [help], [help args], ...)

Will desugar to:

|| Ctx(format!([context], [context args]), format!([help], [help args]))

Note the placement of the ; and ,. They are required and otherwise the macro will not parse.

Example

Assume that we want to run std::fs::read and add context to the error message.

This can be done as follows:

std::fs::read(&path).with_context(ctx!(
  "Could not read the file {path:?}", ;
  "Ensure that the file exists and you have permissions to access it",
));

If one does not want to print a help message this can be easily done by leaving it empty:

std::fs::read(&path).with_context(ctx!(
  "Could not read the file {path:?}", ;
  "",
));