gourd_lib

Module experiment

The setup of an experiment.

Modules

inputs

Dealing with crate::config::UserInputs and InternalInputs

programs

Dealing with crate::config::UserPrograms and InternalPrograms

Structs

Experiment

Describes one experiment.

InternalInput

The internal representation of a crate::config::UserInput

InternalProgram

The internal representation of a crate::config::UserProgram

Metadata

Internally used metadata for inputs/programs

Run

Describes a matching between an algorithm and an input.

RunInput

The input for a Run, exactly as will be passed to the wrapper for execution.

Enums

Environment

An enum to distinguish the run context.

Type Aliases

FieldRef

A string referencing a crate::config::UserProgram, InternalProgram, crate::config::UserInput or InternalInput.