gourd_lib::experiment

Struct Experiment

pub struct Experiment {

    pub seq: usize,
    pub creation_time: DateTime<Local>,
    pub home: PathBuf,
    pub wrapper: String,
    pub inputs: BTreeMap<FieldRef, InternalInput>,
    pub programs: Vec<InternalProgram>,
    pub output_folder: PathBuf,
    pub metrics_folder: PathBuf,
    pub resource_limits: Option<ResourceLimits>,
    pub env: Environment,
    pub num_threads: usize,
    pub labels: BTreeMap<String, Label>,
    pub slurm: Option<SlurmConfig>,
    pub chunks: Vec<Vec<usize>>,
    pub groups: Vec<String>,
    pub runs: Vec<Run>,

}

Describes one experiment.

Fields
seq: usize

The ID of this experiment.

creation_time: DateTime<Local>

The time of creation of the experiment.

home: PathBuf

The directory in which the contents of this experiment reside

wrapper: String

What to call as a std::process::Command to get the wrapper executable.

inputs: BTreeMap<FieldRef, InternalInput>

The inputs for the experiment.

programs: Vec<InternalProgram>

The programs for the experiment.

output_folder: PathBuf

The path to a folder where the experiment output will be stored.

metrics_folder: PathBuf

The path to a folder where the metrics output will be stored.

resource_limits: Option<ResourceLimits>

Global resource limits that will apply to newly created chunks.

env: Environment

Environment of the experiment

num_threads: usize

How many threads to use for local execution

labels: BTreeMap<String, Label>

Labels used in this experiment.

slurm: Option<SlurmConfig>

If running on a SLURM cluster, the job configurations.

chunks: Vec<Vec<usize>>

A mapping of job array task id indices to run ids.

groups: Vec<String>

The input groups present in this experiment.

runs: Vec<Run>

The pairings of program-input for this experiment.

Implementations

impl Experiment

pub fn file(&self) -> PathBuf

Path to the experiment lockfile.

pub fn save_to( &self, folder: &Path, fs: &impl FileOperations, ) -> Result<PathBuf>

Save the experiment to a file with its timestamp.

pub fn save(&self, fs: &impl FileOperations) -> Result<PathBuf>

Save the experiment

pub fn get_label(&self, name: &String) -> Result<Label>

Get the label by name.

pub fn get_program(&self, run: &Run) -> Result<InternalProgram>

Get (a clone of) the InternalProgram used for a given Run.

pub fn program_from_run_id(&self, run_id: usize) -> Result<InternalProgram>

Get (a clone of) the InternalProgram used for a given Run.

pub fn slurm_out(&self, slurm_id: &str) -> Option<PathBuf>

Get the slurm stdout file path for a given run.

pub fn slurm_err(&self, slurm_id: &str) -> Option<PathBuf>

Get the slurm stderr file path for a given run.