gourd_lib::file_system

Trait FileOperations

pub trait FileOperations {
    // Required methods
    fn read_bytes(&self, path: &Path) -> Result<Vec<u8>>;
    fn read_utf8(&self, path: &Path) -> Result<String>;
    fn try_read_toml<T: DeserializeOwned>(&self, path: &Path) -> Result<T>;
    fn try_write_toml<T: Serialize>(&self, path: &Path, data: &T) -> Result<()>;
    fn write_archive<T: Read>(
        &self,
        path: &Path,
        data: Archive<T>,
    ) -> Result<()>;
    fn write_bytes_truncate(&self, path: &Path, bytes: &[u8]) -> Result<()>;
    fn write_utf8_truncate(&self, path: &Path, data: &str) -> Result<()>;
    fn truncate_and_canonicalize(&self, path: &Path) -> Result<PathBuf>;
    fn truncate_and_canonicalize_folder(&self, path: &Path) -> Result<PathBuf>;
    fn set_permissions(&self, path: &Path, perms: u32) -> Result<()>;
    fn canonicalize(&self, path: &Path) -> Result<PathBuf>;
    fn init_git_repository(&self, path: &Path) -> Result<()>;
}

This defines all interactions of gourd with the filesystem.

Required Methods

fn read_bytes(&self, path: &Path) -> Result<Vec<u8>>

Read a file into raw bytes.

fn read_utf8(&self, path: &Path) -> Result<String>

Read a file into a utf8 string.

fn try_read_toml<T: DeserializeOwned>(&self, path: &Path) -> Result<T>

Try to deserialize a toml file into a struture T.

fn try_write_toml<T: Serialize>(&self, path: &Path, data: &T) -> Result<()>

Try to serialize a struct T into a toml file.

fn write_archive<T: Read>(&self, path: &Path, data: Archive<T>) -> Result<()>

Write all files in a .tar directory structure at the provided path.

fn write_bytes_truncate(&self, path: &Path, bytes: &[u8]) -> Result<()>

Write all bytes to a file.

fn write_utf8_truncate(&self, path: &Path, data: &str) -> Result<()>

Write a String to a file.

fn truncate_and_canonicalize(&self, path: &Path) -> Result<PathBuf>

Truncates the file and then runs FileOperations::canonicalize.

fn truncate_and_canonicalize_folder(&self, path: &Path) -> Result<PathBuf>

Truncates the folder and then runs FileOperations::canonicalize.

fn set_permissions(&self, path: &Path, perms: u32) -> Result<()>

Make a file possible to execute.

fn canonicalize(&self, path: &Path) -> Result<PathBuf>

Given a path try to canonicalize it.

This will fail for files that do not exist.

fn init_git_repository(&self, path: &Path) -> Result<()>

Create a new template repository.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.